requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
